requirements.txt
langchain
ipykernel
sentence-transformers
uvicorn
ctransformers
python-box
streamlit
langchain-community
python-dotenv
pypdf
bs4
arxiv
pymupdf
langchain_openai
lxml
langchain-text-splitters
langchain_huggingface
sentence_transformers
faiss-cpu
langchain_chroma
langchain-ollama
langchain_groq
markdown
pipwin
Pygments
SpeechRecognition
pyaudio
googletrans==3.1.0a0
streamlit-lottie
python-docx